Rename native-openc2-command-cisco-ios-acl to native-openc2-command-cisco-ios-acl.json
Cisco ACL configuration demands information that is not part of an OpenC2 command. The native-openc2-command-cisco-ios-acl.json file includes an OpenC2 command with Cisco-relevant information for the configuration of ACLs. This command would be appropriate to be sent to a Native Cisco_IOS interface. Note: That this is just an idea and is not in any case supported by Cisco.